refactor: create `OnboardingInfoButton`
load it dynamically to avoid SSR errors
